dgsm_bot/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dgsm_bot"
version = "0.0.1"
requires-python = ">=3.8"
dependencies = [
    "disnake",
    "ping3",
    "PyYAML",
]